fix prune

+++ commands/prune.js
@@ -7,22 +7,30 @@ module.exports.command = async (message, content, bot, server) => {
 	var ids = content.split(' ');
 	var lastMessageID = message.id;
 	var done = false;
-	var now = parseInt(message.id);
+	var now = (new Date()).getTime();
 	var day = 24 * 60 * 60 * 1000;
+  var count = 0;
 	while (!done) {
-		let messages = await message.channel.fetchMessages({limit: 100, before: lastMessageID});
-		let delMsgs = [];
-		for (var m of messages.values()) {
-			if (now - parseInt(m.id) > day) {
-				done = true;
-				break;
+		let messages = await message.channel.fetchMessages({limit:100,before:lastMessageID});
+    let delMsgs = [];
+    let num = 0;
+    for (var m of messages.values()) {
+      count++;
+      if (++num == 100) {
+        if (now - m.createdAt.getTime() > day) {
+          done = true;
+          break;
+        } else {
+          lastMessageID = m.id;
+        }
 			};
-			lastMessageID = m.id;
 			if (ids.indexOf(m.author.id) != -1) {
 				delMsgs.push(m);
 			}
 		}
-		message.channel.bulkDelete(delMsgs);
+    if (delMsgs.length > 1) {
+		  message.channel.bulkDelete(delMsgs);
+    }
 	}
-	message.channel.send('done!');
+	message.channel.send(`Checked ${count} messages and it's all clear!`);
 };
